refactor(senpi-task): split reconcile-revival under the 250 pure-LOC cap
