+++ packages/senpi-task/src/lifecycle/reconcile-revival.ts
@@ -1,25 +1,26 @@
-import { log } from "@oh-my-opencode/utils"
-
 import type { TaskRecord } from "../state"
-import { isSpawnSpecV1, markRecordLostForReconciliation } from "../state"
-import { delay, nowIso, TERMINAL_STATUSES, type LifecycleContext } from "./context"
-import { destroyResidentTask } from "./destroy"
-import { getLifecycleReattachPorts, type RespawnFailureCode } from "./port"
-import { admitSuspendedBatch, reclaimOrphanedResident } from "./residency"
-import type { ReconcileDeferredReason, ReconcileOutcome } from "./types"
+import { nowIso, TERMINAL_STATUSES, type LifecycleContext } from "./context"
+import {
+  deferred,
+  isClaimHeld,
+  isOrphan,
+  reclaimResident,
+  REVIVABLE_STATUSES,
+  reviveClaimed,
+  type SessionPathResolver,
+  type SuspendedResidency,
+} from "./reconcile-reclamation"
+import { admitSuspendedBatch } from "./residency"
+import type { ReconcileOutcome } from "./types"
 
-const REVIVABLE_STATUSES = new Set(["pending", "running", "completed", "error", "interrupted"])
+export { beginLocalReclamation } from "./reconcile-reclamation"
+
 const SUSPENDED_RESIDENCIES = new Set(["persisted_only", "rpc_detached"])
-const activeLocalReclamations = new Set<string>()
-
-type SuspendedResidency = "persisted_only" | "rpc_detached"
 
 type RevivalCandidate = {
   readonly record: TaskRecord
   readonly priorResidency: SuspendedResidency
 }
-
-type SessionPathResolver = (taskId: string) => string | undefined
 
 export async function reconcileScopedRevival(
   context: LifecycleContext,
@@ -102,192 +103,6 @@ export async function reconcileScopedRevival(
   return outcomes
 }
 
-async function reclaimResident(
-  context: LifecycleContext,
-  observed: TaskRecord,
-  sessionPathFor: SessionPathResolver,
-): Promise<ReconcileOutcome> {
-  const release = beginLocalReclamation(context, observed.task_id)
-  if (release === undefined) return deferred(observed.task_id, "foreign_live_owner")
-  try {
-    return await reclaimResidentExclusive(context, observed, sessionPathFor)
-  } finally {
-    release()
-  }
-}
-
-async function reclaimResidentExclusive(
-  context: LifecycleContext,
-  observed: TaskRecord,
-  sessionPathFor: SessionPathResolver,
-): Promise<ReconcileOutcome> {
-  try {
-    if (reclaimOrphanedResident(context, observed) !== "claimed") {
-      return deferred(observed.task_id, "foreign_live_owner")
-    }
-  } catch {
-    return deferred(observed.task_id, "lock_contended")
-  }
-
-  const claimed = context.store.load(observed.task_id)
-  if (claimed === null || claimed.host_pid !== context.hostPid || claimed.residency_state !== "resident") {
-    return deferred(observed.task_id, "foreign_live_owner")
-  }
-
-  if (claimed.killed === true || claimed.status === "cancelled" || claimed.status === "lost") {
-    await destroyResidentTask(context, claimed.task_id, "reconcile_lost")
-    return {
-      task_id: claimed.task_id,
-      kind: claimed.status === "lost" ? "lost" : "resumed",
-      reason: claimed.killed === true ? "killed orphan disposed" : `${claimed.status} orphan disposed`,
-    }
-  }
-  if (!REVIVABLE_STATUSES.has(claimed.status)) {
-    await destroyResidentTask(context, claimed.task_id, "reconcile_lost")
-    return { task_id: claimed.task_id, kind: "resumed", reason: "non-revivable orphan disposed" }
-  }
-
-  const sessionPath = sessionPathFor(claimed.task_id)
-  if (TERMINAL_STATUSES.has(claimed.status) && sessionPath === undefined) {
-    context.store.transition(claimed.task_id, { type: "dispose", timestamp: nowIso(context) })
-    return {
-      task_id: claimed.task_id,
-      kind: "resumed",
-      reason: "terminal without transcript disposed; persisted result preserved",
-    }
-  }
-
-  const rollbackResidency: SuspendedResidency = claimed.execution_mode === "process" ? "rpc_detached" : "persisted_only"
-  if (context.config.reattach_on_reconcile === false) {
-    return rollbackOrDeferred(context, claimed.task_id, rollbackResidency, "reattach_disabled")
-  }
-  return reviveClaimed(context, claimed, rollbackResidency, sessionPath)
-}
-
-async function reviveClaimed(
-  context: LifecycleContext,
-  claimed: TaskRecord,
-  rollbackResidency: SuspendedResidency,
-  sessionPath: string | undefined,
-): Promise<ReconcileOutcome> {
-  const fresh = context.store.load(claimed.task_id)
-  if (!isClaimHeld(context, fresh, claimed.parent_session_id) || fresh.killed === true || !REVIVABLE_STATUSES.has(fresh.status)) {
-    return rollbackOrDeferred(context, claimed.task_id, rollbackResidency, "foreign_live_owner")
-  }
-
-  if (fresh.execution_mode === "process" && fresh.pid !== undefined) {
-    const terminated = await terminateOldRpc(context, fresh)
-    if (!terminated) {
-      return rollbackOrDeferred(context, fresh.task_id, rollbackResidency, "session_unavailable")
-    }
-  }
-
-  if (sessionPath === undefined && !isSpawnSpecV1Record(fresh)) {
-    if (TERMINAL_STATUSES.has(fresh.status)) {
-      context.store.transition(fresh.task_id, { type: "dispose", timestamp: nowIso(context) })
-      return {
-        task_id: fresh.task_id,
-        kind: "resumed",
-        reason: "terminal without transcript disposed; persisted result preserved",
-      }
-    }
-    await markLost(context, fresh, "record has neither a session transcript nor a persisted v1 spawn spec")
-    return { task_id: fresh.task_id, kind: "lost", reason: "spawn spec unavailable" }
-  }
-
-  const ports = context.reattachPorts ?? getLifecycleReattachPorts(context.store)
-  if (ports === undefined) {
-    await markLost(context, fresh, "reattach ports unavailable")
-    return { task_id: fresh.task_id, kind: "lost", reason: "reattach ports unavailable" }
-  }
-
-  const respawned = await ports.respawn(fresh, sessionPath)
-  if (!respawned.ok) {
-    if (respawned.disposition === "retryable") {
-      return rollbackOrDeferred(context, fresh.task_id, rollbackResidency, deferredCode(respawned.code))
-    }
-    if (TERMINAL_STATUSES.has(fresh.status)) {
-      context.store.transition(fresh.task_id, { type: "dispose", timestamp: nowIso(context) })
-      return { task_id: fresh.task_id, kind: "resumed", reason: respawned.reason }
-    }
-    await markLost(context, fresh, `reattach failed: ${respawned.reason}`)
-    return { task_id: fresh.task_id, kind: "lost", reason: respawned.reason }
-  }
-
-  const reattached = await ports.reattach(fresh, respawned.handle)
-  if (!reattached.ok) {
-    if (reattached.kind === "already_attached") {
-      return { task_id: fresh.task_id, kind: "resumed", reason: reattached.reason }
-    }
-    return rollbackOrDeferred(context, fresh.task_id, rollbackResidency, "session_unavailable")
-  }
-  context.store.appendEvent(fresh.task_id, {
-    type: "reconcile_reattached",
-    payload: sessionPath === undefined ? { fresh_launch: true } : { session_path: sessionPath },
-  })
-  return { task_id: fresh.task_id, kind: "resumed", reason: "respawned and reattached" }
-}
-
-async function terminateOldRpc(context: LifecycleContext, record: TaskRecord): Promise<boolean> {
-  const pid = record.pid
-  if (pid === undefined || !context.signaller.isAlive(pid)) return true
-  context.signaller.signal(pid, "SIGTERM")
-  context.store.appendEvent(record.task_id, { type: "reconcile_terminated", payload: { pid, signal: "SIGTERM" } })
-  await delay(context.orphanKillDelayMs)
-  if (context.signaller.isAlive(pid)) {
-    context.signaller.signal(pid, "SIGKILL")
-    context.store.appendEvent(record.task_id, { type: "reconcile_terminated", payload: { pid, signal: "SIGKILL" } })
-  }
-  return !context.signaller.isAlive(pid)
-}
-
-function rollbackOrDeferred(
-  context: LifecycleContext,
-  taskId: string,
-  residency: SuspendedResidency,
-  successReason: ReconcileDeferredReason,
-): ReconcileOutcome {
-  return rollbackClaim(context, taskId, residency)
-    ? deferred(taskId, successReason)
-    : deferred(taskId, "rollback_failed")
-}
-
-function rollbackClaim(context: LifecycleContext, taskId: string, residency: SuspendedResidency): boolean {
-  try {
-    context.store.mutate(taskId, (fresh) => {
-      if (fresh.host_pid !== context.hostPid || fresh.residency_state !== "resident") return fresh
-      const { host_pid: _hostPid, ...withoutHost } = fresh
-      if (residency === "rpc_detached") {
-        return { ...withoutHost, residency_state: residency, updated_at: nowIso(context) }
-      }
-      const { pid: _pid, ...withoutPid } = withoutHost
-      return { ...withoutPid, residency_state: residency, updated_at: nowIso(context) }
-    })
-    return true
-  } catch (error) {
-    log("senpi-task reconcile ownership rollback failed", { taskId, residency, error: String(error) })
-    return false
-  }
-}
-
-async function markLost(context: LifecycleContext, record: TaskRecord, message: string): Promise<void> {
-  let applied = false
-  context.store.mutate(record.task_id, (fresh) => {
-    if (fresh.host_pid !== context.hostPid || fresh.residency_state !== "resident") return fresh
-    const result = markRecordLostForReconciliation(fresh, {
-      timestamp: nowIso(context),
-      error_message: message,
-      updateReason: fresh.status === "lost",
-    })
-    if (!result.applied) return fresh
-    applied = true
-    return result.record
-  })
-  if (!applied) return
-  context.store.appendEvent(record.task_id, { type: "reconcile_lost", payload: { reason: message } })
-  await destroyResidentTask(context, record.task_id, "reconcile_lost")
-}
-
 function disposeSuspendedTerminalWithoutTranscript(
   context: LifecycleContext,
   observed: TaskRecord,
@@ -314,43 +129,6 @@ function suspendedCandidates(context: LifecycleContext, parentSessionId: string)
   })
 }
 
-function isOrphan(context: LifecycleContext, record: TaskRecord): boolean {
-  if (record.host_pid === context.hostPid) return !hasLiveHandle(context, record.task_id)
-  return record.host_pid === undefined || !context.signaller.isAlive(record.host_pid)
-}
-
-function hasLiveHandle(context: LifecycleContext, taskId: string): boolean {
-  return context.registry.get(taskId) !== undefined || context.registry.entries().some((handle) => handle.task_id === taskId)
-}
-
-function isClaimHeld(
-  context: LifecycleContext,
-  record: TaskRecord | null,
-  parentSessionId: string,
-): record is TaskRecord {
-  return record !== null && record.parent_session_id === parentSessionId &&
-    record.residency_state === "resident" && record.host_pid === context.hostPid
-}
-
 function isSuspended(record: TaskRecord): record is TaskRecord & { readonly residency_state: SuspendedResidency } {
   return record.residency_state === "persisted_only" || record.residency_state === "rpc_detached"
-}
-
-function isSpawnSpecV1Record(record: TaskRecord): boolean {
-  return record.spawn_spec !== undefined && isSpawnSpecV1(record.spawn_spec)
-}
-
-function deferredCode(code: RespawnFailureCode): ReconcileDeferredReason {
-  return code === "respawn_failed" ? "session_unavailable" : code
-}
-
-export function beginLocalReclamation(context: LifecycleContext, taskId: string): (() => void) | undefined {
-  const key = `${context.store.stateDir}\u0000${taskId}`
-  if (activeLocalReclamations.has(key)) return undefined
-  activeLocalReclamations.add(key)
-  return () => activeLocalReclamations.delete(key)
-}
-
-function deferred(taskId: string, reason: ReconcileDeferredReason): ReconcileOutcome {
-  return { task_id: taskId, kind: "deferred", reason }
 }
